level extraction + loading in Game.jsx
